Skip ONNX optimization when it fails
During ONNX export, model optimization is performed. Currently, if
optimization fails, the export also fails

This PR allows export to succeed even when the optimization fails.
This is useful to unblock users and allow them to inspect the ONNX model
actquired through the export process

A warning with a link to report the bug is printed to let user know
about the issue